LayoutTests/http/tests/cookies/document-cookie-multiple-cookies-expected.txt

Tests trying to set multiple cookies at once using document.cookie

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.cookie is ""
PASS document.cookie is "samesite-unspecified=0"
PASS successfullyParsed is true

TEST COMPLETE
